Refactor: 패키지 경로 수정(inventory, reservation 분리)
