+++ com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/application/dto/InventoryQueryResponseDto.java
@@ -1,6 +1,6 @@
 package com.bkmarriott.reservationservice.reservation.application.dto;
 
-import com.bkmarriott.reservationservice.reservation.domain.vo.RoomType;
+import com.bkmarriott.reservationservice.reservation.domain.vo.inventory.RoomType;
 import com.bkmarriott.reservationservice.reservation.infrastructure.persistence.entity.RoomEntityType;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
