+++ com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/application/outputport/InventoryQueryOutputPort.java
@@ -1,9 +1,9 @@
 package com.bkmarriott.reservationservice.reservation.application.outputport;
 
-import com.bkmarriott.reservationservice.reservation.application.dto.InventoryQueryRequestDto;
-import com.bkmarriott.reservationservice.reservation.application.dto.InventoryQueryResponseDto;
+import com.bkmarriott.reservationservice.reservation.application.dto.inventory.InventoryQueryRequestDto;
+import com.bkmarriott.reservationservice.reservation.application.dto.inventory.InventoryQueryResponseDto;
 import com.bkmarriott.reservationservice.reservation.domain.Inventory;
-import com.bkmarriott.reservationservice.reservation.domain.vo.RoomType;
+import com.bkmarriott.reservationservice.reservation.domain.vo.inventory.RoomType;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
