+++ com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/application/service/InventoryService.java
@@ -11,7 +11,7 @@ import com.bkmarriott.reservationservice.reservation.domain.Reservation;
 import com.bkmarriott.reservationservice.reservation.domain.vo.inventory.RoomType;
 import com.bkmarriott.reservationservice.reservation.domain.vo.reservation.ReservationForCreate;
 import com.bkmarriott.reservationservice.reservation.domain.vo.reservation.ReservationStatus;
-import com.bkmarriott.reservationservice.reservation.presentation.rest.dto.inventory.query.InventoryQuery.Response;
+import com.bkmarriott.reservationservice.reservation.presentation.rest.dto.inventory.InventoryQuery.Response;
 import com.bkmarriott.reservationservice.reservation.application.exception.ResourceNotFoundException;
 import java.time.LocalDate;
 import java.util.List;
