+++ com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/infrastructure/persistence/adapter/InventoryQueryAdaptor.java
@@ -1,10 +1,10 @@
 package com.bkmarriott.reservationservice.reservation.infrastructure.persistence.adapter;
 
-import com.bkmarriott.reservationservice.reservation.application.dto.InventoryQueryRequestDto;
-import com.bkmarriott.reservationservice.reservation.application.dto.InventoryQueryResponseDto;
+import com.bkmarriott.reservationservice.reservation.application.dto.inventory.InventoryQueryRequestDto;
+import com.bkmarriott.reservationservice.reservation.application.dto.inventory.InventoryQueryResponseDto;
 import com.bkmarriott.reservationservice.reservation.application.outputport.InventoryQueryOutputPort;
 import com.bkmarriott.reservationservice.reservation.domain.Inventory;
-import com.bkmarriott.reservationservice.reservation.domain.vo.RoomType;
+import com.bkmarriott.reservationservice.reservation.domain.vo.inventory.RoomType;
 import com.bkmarriott.reservationservice.reservation.infrastructure.persistence.entity.RoomTypeInventoryEntity;
 import com.bkmarriott.reservationservice.reservation.infrastructure.persistence.entity.RoomTypeInventoryId;
 import com.bkmarriott.reservationservice.reservation.infrastructure.persistence.repository.InventoryQueryDslRepository;
