--- com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/infrastructure/persistence/config/PersistenceConfig.java
@@ -1,10 +1,0 @@
-package com.bkmarriott.reservationservice.reservation.infrastructure.persistence.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class PersistenceConfig {
-
-}
