+++ com.bkmarriott.reservation/src/main/java/com/bkmarriott/reservationservice/reservation/presentation/rest/controller/inventory/InventoryCommandController.java
@@ -3,7 +3,7 @@ package com.bkmarriott.reservationservice.reservation.presentation.rest.controll
 import static com.bkmarriott.reservationservice.reservation.presentation.config.HttpHeaderConstants.HEADER_USER_ID;
 
 import com.bkmarriott.reservationservice.reservation.application.service.InventoryService;
-import com.bkmarriott.reservationservice.reservation.presentation.rest.dto.inventory.command.InventoryModification.Response;
+import com.bkmarriott.reservationservice.reservation.presentation.rest.dto.inventory.InventoryModification.Response;
 import com.bkmarriott.reservationservice.reservation.presentation.rest.util.reponse.ApiResponse;
 import com.bkmarriott.reservationservice.reservation.presentation.rest.util.reponse.ApiResponse.Success;
 import java.util.List;
